Move 'ShutdownMap' into the 'provider' module.
